LPD-88211 Move dropdown actions to SectionDisplayContextHelper
Different display contexts were using slightly different dropdown
lists (in some cases in different sort orders). Move everything to the
display context helper and make everything consistent.

Some actions are still left as special cases ("find-and-replace",
"download-folder"), as they don't apply to all sections.
